Rename packages in AWS module
consolidate build scripts with convention plugin
report jacoco on PR
