+++ .gitignore
@@ -2,3 +2,4 @@
 .idea
 build/
 .obsidian/
+.kotlin/
