+++ buildSrc/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    kotlin("jvm") version ("2.0.0")
+    `kotlin-dsl`
+}
+repositories {
+    gradlePluginPortal()
+}
