+++ buildSrc/settings.gradle.kts
@@ -1,0 +1,2 @@
+
+rootProject.name="funk-conventions"
