+++ spring-funk-aws/src/main/kotlin/com/github/wakingrufus/funk/aws/s3/S3ClientConfigurationProperties.kt
@@ -1,4 +1,4 @@
-package com.github.wakingrufus.springdsl.aws.s3
+package com.github.wakingrufus.funk.aws.s3
 
 import java.time.Duration
