+++ spring-funk-aws/src/noMicrometer/kotlin/S3ClientInitializerTest.kt
@@ -1,5 +1,5 @@
-import com.github.wakingrufus.springdsl.aws.aws
-import com.github.wakingrufus.springdsl.aws.s3.S3ClientInitializer
+import com.github.wakingrufus.funk.aws.aws
+import com.github.wakingrufus.funk.aws.s3.S3ClientInitializer
 import com.github.wakingrufus.springdsl.test.testDslApplication
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
