+++ spring-funk-core/build.gradle.kts
@@ -1,28 +1,8 @@
 plugins {
-    id("java-library")
     kotlin("jvm")
-    `jvm-test-suite`
-    jacoco
-}
-
-repositories {
-    mavenCentral()
+    funk.library
 }
 
 dependencies {
     testImplementation(libs.assertj)
-}
-
-java {
-    toolchain {
-        languageVersion.set(JavaLanguageVersion.of(21))
-    }
-}
-
-testing {
-    suites {
-        val test by getting(JvmTestSuite::class) {
-            useJUnitJupiter()
-        }
-    }
 }
